+++ src/util/util.go
@@ -2,20 +2,20 @@ package util
 
 import "strings"
 
-func ConvertCol(col_square byte) int {
+func ConvertCol(col_square byte) uint8 {
 	col_square = strings.ToUpper(string(col_square))[0]
-	return int(col_square - 'A')
+	return uint8(col_square - 'A')
 }
 
-func ConvertRow(row_square byte) int {
-	return 5 - int((row_square - '1'))
+func ConvertRow(row_square byte) uint8 {
+	return 5 - (row_square - '1')
 }
 
 func ConvertColBack(col_index int) byte {
 	return byte(col_index + 'A')
 }
 
-func ConvertSquare(square string) int {
+func ConvertSquare(square string) uint8 {
 	col, row := square[0], square[1]
 	colIndex, rowIndex := ConvertCol(col), ConvertRow(row)
 	return colIndex + rowIndex*7
